yaw rate interface
git-svn-id: http://svn.lee.org/swarm/trunk@863 47a5675e-1c2e-0410-904b-cc838fe3ce0b

+++ spu/dispatcher/imuutils.h
@@ -29,4 +29,4 @@ void logDriveDataString(struct swarmMotorData *motorData, char *motorDataString)
 void logSteerDataString(struct swarmMotorData *motorData, char *motorDataString);
 int parseDriveMsg(char *driveBuf, struct swarmMotorData *motData);
 void initYawSensor(void);
-int getYawRate(void);
+double getYawRate(void);
